# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Projecte Allotjaments Sol",
        "description": "REST API. Projecte Allotjaments Sol. DAW Client i servidor. Per utilitzar les cridades que necesiten token primer\n  s'haura de fer login amb un usuari per obtenir el token i despres utilitzar aquest token a la part on posa Authorize. Per\n  mes informacio, obtenir un usuari o altres peticions utilitzar el link de Contact Equip Sol o mandar un mail a jaumefullana@paucasesnovescifp.cat",
        "contact": {
            "name": "Equip Sol",
            "email": "jaumefullana@paucasesnovescifp.cat"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:81/allotjamentssol/public/"
        }
    ],
    "paths": {
        "/api/allotjament": {
            "get": {
                "tags": [
                    "Allotjaments"
                ],
                "summary": "Mostrar tots els allotjaments.",
                "operationId": "1add0ed8eea62f5417aa47f81e1a0b6a",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Status": {
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "Result": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Allotjaments"
                ],
                "summary": "Modifica un allotjament",
                "description": "Modifica un allotjament.",
                "operationId": "6a0ab1ad7937df8328c58ecc5162b396",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "ID": {
                                        "type": "number",
                                        "format": "number",
                                        "example": "1"
                                    },
                                    "NumeroRegistre": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "A1112223334444"
                                    },
                                    "NomComercial": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Can Tomeu Llarg"
                                    },
                                    "Direccio": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "C\\ Gran 35"
                                    },
                                    "NumeroLlits": {
                                        "type": "number",
                                        "format": "number",
                                        "example": "3"
                                    },
                                    "NumeroPersones": {
                                        "type": "number",
                                        "format": "number",
                                        "example": "3"
                                    },
                                    "NumeroHabitacions": {
                                        "type": "number",
                                        "format": "number",
                                        "example": "7"
                                    },
                                    "NumeroBanys": {
                                        "type": "number",
                                        "format": "number",
                                        "example": "2"
                                    },
                                    "MunicipisID": {
                                        "type": "number",
                                        "format": "number",
                                        "example": "7"
                                    },
                                    "TipusVacancesID": {
                                        "type": "number",
                                        "format": "number",
                                        "example": "9"
                                    },
                                    "TipusAllotjamentsID": {
                                        "type": "number",
                                        "format": "number",
                                        "example": "2"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Status": {
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "Result": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Status": {
                                            "type": "string",
                                            "example": "Error"
                                        },
                                        "Result": {
                                            "type": "string",
                                            "example": "Informacio de l'error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Allotjaments"
                ],
                "summary": "Crea un nou allotjament",
                "description": "Crea un nou allotjament.",
                "operationId": "f79a7117b3ed9e9664bc4152f15015b3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "NumeroRegistre": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "A1112223334444"
                                    },
                                    "NomComercial": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Can Tomeu Llarg"
                                    },
                                    "Direccio": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "C\\ Gran 35"
                                    },
                                    "NumeroLlits": {
                                        "type": "number",
                                        "format": "number",
                                        "example": "3"
                                    },
                                    "NumeroPersones": {
                                        "type": "number",
                                        "format": "number",
                                        "example": "3"
                                    },
                                    "NumeroHabitacions": {
                                        "type": "number",
                                        "format": "number",
                                        "example": "7"
                                    },
                                    "NumeroBanys": {
                                        "type": "number",
                                        "format": "number",
                                        "example": "2"
                                    },
                                    "MunicipisID": {
                                        "type": "number",
                                        "format": "number",
                                        "example": "7"
                                    },
                                    "TipusVacancesID": {
                                        "type": "number",
                                        "format": "number",
                                        "example": "9"
                                    },
                                    "TipusAllotjamentsID": {
                                        "type": "number",
                                        "format": "number",
                                        "example": "2"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Status": {
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "Result": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Status": {
                                            "type": "string",
                                            "example": "Error"
                                        },
                                        "Result": {
                                            "type": "string",
                                            "example": "Informacio de l'error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Allotjaments"
                ],
                "summary": "Esborra un allotjament",
                "description": "Esborra un allotjament.",
                "operationId": "458feb8dec4d7e990097f89a0055ab03",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "ID": {
                                        "type": "number",
                                        "format": "number",
                                        "example": "5"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Status": {
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "Result": {
                                            "type": "boolean",
                                            "example": "true"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Status": {
                                            "type": "string",
                                            "example": "Error"
                                        },
                                        "Result": {
                                            "type": "string",
                                            "example": "Informacio de l'error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/allotjament/{id}": {
            "get": {
                "tags": [
                    "Allotjaments"
                ],
                "summary": "Mostrar l'allotjament al qual correspon la id passada per la url",
                "operationId": "9ed73a121b82970eff4eb19cbbb59a19",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id de l'allotjament",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Status": {
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "Result": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "scheme": "bearer"
            }
        }
    }
}